Python3 fix
git-svn-id: https://svn.osgeo.org/gdal/trunk@30031 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/alg/sieve.py
@@ -211,7 +211,7 @@ def sieve_6():
     ds = gdal.GetDriverByName('MEM').Create('', size+1, size)
     ar = numpy.zeros((size,size+1), dtype = numpy.uint8)
     for i in range(size):
-        for j in range(size/3):
+        for j in range(int(size/3)):
             ar[i][size+1-1-i-1-3*j] = 255
             ar[i][size+1-1-i-3*j] = 255
         ar[i][0] = 255
